requirements.txt
# Web Framework
flask
flask-cors

# Data Processing
numpy
pandas
pyspark

# NLP and Machine Learning
nltk
scikit-learn

# Resume Parsing
python-docx
PyPDF2
pdfplumber

# Utils
python-dotenv
requests

# HTML to PDF Resume Generation
Jinja2
WeasyPrint

reportlab
certifi 

pymongo
PyJWT==2.8.0
werkzeug==2.3.7

thefuzz[speedup]

